feat: migrar módulo notificaciones a Drizzle + fix count(*) como string
getNotificaciones/getUnreadCount/marcarTodasLeidas ahora usan Drizzle. La
autenticación de marcarTodasLeidasAction sigue en Supabase Auth (fase
aparte, todavía sin migrar) — solo se movió la query de datos.

De paso corrige un bug real detectado con un test: count(*) via Drizzle/
postgres.js devuelve un string en runtime aunque se tipee sql<number>
(bigint de Postgres → string en JS por precisión). Afectaba también al
count de getClientes, ya migrado. Se soluciona casteando count(*)::int
en el SQL en vez de confiar en el tipo de TypeScript.

+++ src/modules/clientes/services/clientes.service.ts
@@ -35,7 +35,7 @@ export async function getClientes(search?: string, page = 1) {
 
   const [data, countResult] = await Promise.all([
     db.select(clienteSelect).from(clientes).where(where).orderBy(asc(clientes.nombre)).limit(PAGE_SIZE).offset(from),
-    db.select({ total: sql<number>`count(*)` }).from(clientes).where(where),
+    db.select({ total: sql<number>`count(*)::int` }).from(clientes).where(where),
   ])
 
   return { data, count: countResult[0].total, pageSize: PAGE_SIZE }
